+++ examples/__init__.py
@@ -1,0 +1,1 @@
+"""Example packages for neo4j-python-pandas-py2neo-v3."""